Added a sub to transpose the image array
Imager::Montage takes an array of images and generates a grid with given
cols and rows. It fills the grid column first e.g.

Given the array:
(a,b,c,d,e,f)

asking for a rows => 2 cols => 3 grid would give:
a c e
b d f

If we call transpose on the array first it will rearrange the elements
in such a way that we get
a b c
d e f

as the output from Imager::Montage. I am not sure if this should be done
by default or based on a command line switch. Keeping it commented out
till I make up my mind.